+++ packages/core/src/index.ts
@@ -8,6 +8,7 @@
 
 import './styles';
 
+export { default as AnnotationType } from './annotations/AnnotationType';
 export { default as Button } from './components/Button';
 export { default as Menu } from './components/Menu';
 export { default as MenuDivider } from './components/MenuDivider';